// frontend/src/api/generated/expense-trackers/expense-trackers.zod.ts
export const postApiExpenseTrackersBodyAmountExclusiveMin = 0;
export const postApiExpenseTrackersBodyAmountMax = 999999.99;

export const postApiExpenseTrackersBodyCategoryMax = 50;

export const postApiExpenseTrackersBodyDescriptionMax = 500;

export const PostApiExpenseTrackersBody = zod.object({
  amount: zod
    .number()
    .gt(postApiExpenseTrackersBodyAmountExclusiveMin)
    .max(postApiExpenseTrackersBodyAmountMax),
  category: zod.string().min(1).max(postApiExpenseTrackersBodyCategoryMax),
  description: zod
    .string()
    .max(postApiExpenseTrackersBodyDescriptionMax)
    .nullish(),
  date: zod.iso.datetime({}).min(1),
});